requirements.txt
appdirs==1.4.4
asgiref==3.3.1
autopep8==1.5.4
awscli==1.19.9
boto3==1.17.2
botocore==1.20.9
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.3
cryptography==3.4.6
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
Dijkstar==2.5.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.7
django-allauth==0.44.0
django-crispy-forms==1.11.1
django-widget-tweaks==1.4.8
docutils==0.15.2
filelock==3.0.12
google-auth==1.25.0
google-auth-oauthlib==0.4.2
graphviz==0.15
gspread==3.6.0
gunicorn==20.1.0
httplib2==0.18.1
idna==2.10
jmespath==0.10.0
kiwisolver==1.3.1
matplotlib==3.3.3
networkx==2.5
numpy==1.19.4
oauth2client==4.1.3
oauthlib==3.1.0
Pillow==8.0.1
pipenv==2020.11.15
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
python-louvain==0.15
python3-openid==3.2.0
pytube==10.4.1
pytz==2021.1
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.5
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.3
virtualenv==20.4.2
virtualenv-clone==0.5.4
whitenoise==5.2.0
